✨ Enroll deck feature
