+++ src/app-container.ts
@@ -40,6 +40,7 @@ import { Server as ApolloServer } from '@api/apollo/apollo.server';
 import { QueryHandler } from '@app/processing/query-handler';
 import { QueryBus } from '@app/processing/query-bus';
 import { GetDeckByIdQueryHandler } from '@app/decks/get-deck-by-id/get-deck-by-id.query-handler';
+import { EnrollDeckCommandHandler } from '@app/decks/enroll-deck/enroll-deck.command-handler';
 
 const registerAsArray = <T>(resolvers: Awilix.Resolver<T>[]): Awilix.Resolver<T[]> => ({
   resolve: (container: Awilix.AwilixContainer) => resolvers.map((r) => container.build(r)),
@@ -96,6 +97,7 @@ export const createAppContainer = async (): Promise<Awilix.AwilixContainer> => {
       Awilix.asClass(RemoveCardCommandHandler).singleton(),
       Awilix.asClass(DeleteDeckCommandHandler).singleton(),
       Awilix.asClass(PublishDeckCommandHandler).singleton(),
+      Awilix.asClass(EnrollDeckCommandHandler).singleton(),
     ]),
   });
