+++ src/errors/invariant.error.ts
@@ -1,0 +1,7 @@
+import { AppError } from './app.error';
+
+export class InvariantError extends AppError {
+  constructor(message: string) {
+    super(message, 'InvariantError');
+  }
+}
